modernize `huggingface_hub` usage (#89)
This PR adjusts the interface of `from_pretrained` and `save_pretrained`
to allow for the parameters of `huggingface_hub` version
[0.23.4](https://github.com/huggingface/huggingface_hub/releases/tag/v0.23.4)
(which is our current **min** version) for `hf_hub_download` and
`upload_folder`, respectively.

This also mitigates the warning: 
```
/home/arbi01/.cache/pypoetry/virtualenvs/pie-core-FdWNV-Ov-py3.10/lib/python3.10/site-packages/huggingface_hub/file_download.py:1132: 
FutureWarning: `resume_download` is deprecated and will be removed in version 1.0.0. Downloads always resume when possible. If you want to force a new download, use `force_download=True`.
```

breaking, because this removes deprecated parameters from
`_from_pretrained` (e.g. `resume_download`)