`Auto.from_config`: allow to pass the class name per keyword argument, even if no respective key is in the config (#95)
This is relevant to use `AutoAnnotationPipeline` with "old"
models/configs that do not have the key `pipeline_type`. With this PR,
it simply can be provided as argument to
`AutoAnnotationPipeline.from_pretrained`. This allows to use former
method in any inference setup and thus keep this *PIE*-generic, i.e.,
independent of *PyTorch-IE*.

Example:
```
pipeline = AutoAnnotationPipeline.from_pretrained("pie/example-ner-spanclf-conll03", pipeline_type="pytorch-ie")
```